STAF-34: fix IconButton/ Header components
