+++ src/core/components/HeaderBar/HeaderBar.stories.tsx
@@ -1,0 +1,18 @@
+import { ComponentStory, ComponentMeta } from '@storybook/react';
+import { HeaderBar } from './HeaderBar';
+
+export default {
+  title: 'HeaderBar',
+  component: HeaderBar,
+} as ComponentMeta<typeof HeaderBar>;
+
+const Template: ComponentStory<typeof HeaderBar> = (args) => {
+  const HeaderBarProps = {
+    ...args,
+    title: 'Thai Food - Thai Cook',
+  };
+
+  return <HeaderBar {...HeaderBarProps} />;
+};
+
+export const Primary = Template.bind({});
