+++ src/core/components/HeaderBar/index.ts
@@ -1,0 +1,1 @@
+export { HeaderBar } from './HeaderBar';
